[ENH] Move experimental surface classes to the stable API (#4723)
