+++ doc/modules/experimental.rst
@@ -47,11 +47,6 @@ Classes
    :toctree: generated/
    :template: class.rst
 
-   FileMesh
-   InMemoryMesh
-   Mesh
-   PolyMesh
-   SurfaceImage
    SurfaceLabelsMasker
 
 Functions
