+++ examples/01_plotting/plot_surf_stat_map.py
@@ -35,11 +35,11 @@ See :ref:`plotting` for more details on plotting tools.
 # NKI resting state data from nilearn
 from nilearn.datasets import fetch_atlas_surf_destrieux
 from nilearn.experimental.surface import (
-    SurfaceImage,
     fetch_nki,
     load_fsaverage,
     load_fsaverage_data,
 )
+from nilearn.surface import SurfaceImage
 
 nki_dataset = fetch_nki(n_subjects=1)
