+++ examples/07_advanced/plot_surface_bids_analysis.py
@@ -71,7 +71,8 @@ models, run_imgs, events, confounds = first_level_from_bids(
 # Note that those correspond spatially, as they are both in :term:`MNI` space.
 from pathlib import Path
 
-from nilearn.experimental.surface import SurfaceImage, load_fsaverage
+from nilearn.experimental.surface import load_fsaverage
+from nilearn.surface import SurfaceImage
 
 fsaverage5 = load_fsaverage()
 
@@ -125,9 +126,8 @@ z_val_right = norm.isf(pval_right)
 
 # %%
 # Plot the resulting maps, at first on the left hemisphere.
-from nilearn.experimental.plotting import plot_surf_stat_map
 from nilearn.experimental.surface import load_fsaverage_data
-from nilearn.plotting import show
+from nilearn.plotting import plot_surf_stat_map, show
 
 fsaverage_data = load_fsaverage_data(data_type="sulcal")
