+++ examples/07_advanced/plot_surface_image_and_maskers.py
@@ -15,7 +15,7 @@ from nilearn.experimental.surface import fetch_nki
 from nilearn.maskers import SurfaceMasker
 from nilearn.plotting import plot_matrix, plot_surf, plot_surf_roi, show
 
-img = fetch_nki()[0]
+img = fetch_nki(n_subjects=1)[0]
 print(f"NKI image: {img}")
 
 masker = SurfaceMasker()
@@ -62,11 +62,11 @@ show()
 from nilearn.connectome import ConnectivityMeasure
 from nilearn.datasets import fetch_atlas_surf_destrieux
 from nilearn.experimental.surface import (
-    SurfaceImage,
     SurfaceLabelsMasker,
     load_fsaverage,
     load_fsaverage_data,
 )
+from nilearn.surface import SurfaceImage
 
 fsaverage = load_fsaverage("fsaverage5")
 destrieux = fetch_atlas_surf_destrieux()
@@ -82,7 +82,7 @@ label_names = [x.decode("utf-8") for x in destrieux.labels]
 # for our plots we will be using the fsaverage sulcal data as background map
 fsaverage_sulcal = load_fsaverage_data(data_type="sulcal")
 
-img = fetch_nki()[0]
+img = fetch_nki(n_subjects=1)[0]
 print(f"NKI image: {img}")
 
 fsaverage = load_fsaverage("fsaverage5")
@@ -144,7 +144,7 @@ import numpy as np
 
 from nilearn.decoding import Decoder
 
-img = fetch_nki()[0]
+img = fetch_nki(n_subjects=1)[0]
 
 # create some random labels
 rng = np.random.RandomState(0)
@@ -175,7 +175,7 @@ show()
 # ---------------------------------------
 from sklearn import feature_selection, linear_model, pipeline, preprocessing
 
-img = fetch_nki()[0]
+img = fetch_nki(n_subjects=1)[0]
 y = rng.normal(size=img.shape[0])
 
 decoder = pipeline.make_pipeline(
