+++ nilearn/conftest.py
@@ -15,7 +15,7 @@ from nilearn.datasets.tests._testing import (
     request_mocker,  # noqa: F401
     temp_nilearn_data_dir,  # noqa: F401
 )
-from nilearn.experimental.surface import (
+from nilearn.surface import (
     InMemoryMesh,
     PolyMesh,
     SurfaceImage,
