+++ nilearn/experimental/surface/__init__.py
@@ -8,17 +8,17 @@ from nilearn.experimental.surface._datasets import (
 from nilearn.experimental.surface._surface_image import (
     FileMesh,
     InMemoryMesh,
-    Mesh,
     PolyData,
     PolyMesh,
     SurfaceImage,
+    SurfaceMesh,
 )
 from nilearn.experimental.surface.maskers import SurfaceLabelsMasker
 
 __all__ = [
     "FileMesh",
     "InMemoryMesh",
-    "Mesh",
+    "SurfaceMesh",
     "PolyMesh",
     "PolyData",
     "SurfaceImage",
