+++ nilearn/experimental/surface/maskers.py
@@ -10,12 +10,12 @@ from sklearn.base import BaseEstimator
 
 from nilearn._utils import _constrained_layout_kwargs
 from nilearn._utils.helpers import is_matplotlib_installed
-from nilearn.experimental.surface._surface_image import SurfaceImage
 from nilearn.maskers._utils import (
     check_same_n_vertices,
     compute_mean_surface_image,
     get_min_max_surface_image,
 )
+from nilearn.surface import SurfaceImage
 
 
 class SurfaceLabelsMasker(BaseEstimator):
@@ -266,7 +266,7 @@ class SurfaceLabelsMasker(BaseEstimator):
     def _create_figure_for_report(self):
         import matplotlib.pyplot as plt
 
-        from nilearn.experimental import plotting
+        from nilearn import plotting
 
         labels_img = self._reporting_data["labels_image"]
