+++ nilearn/plotting/tests/test_surf_plotting.py
@@ -13,7 +13,6 @@ from numpy.testing import assert_array_equal
 from nilearn._utils.helpers import is_kaleido_installed, is_plotly_installed
 from nilearn.conftest import _rng
 from nilearn.datasets import fetch_surf_fsaverage
-from nilearn.experimental.surface import InMemoryMesh, SurfaceImage
 from nilearn.plotting._utils import check_surface_plotting_inputs
 from nilearn.plotting.displays import PlotlySurfaceFigure, SurfaceFigure
 from nilearn.plotting.surf_plotting import (
@@ -29,7 +28,12 @@ from nilearn.plotting.surf_plotting import (
     plot_surf_roi,
     plot_surf_stat_map,
 )
-from nilearn.surface import load_surf_data, load_surf_mesh
+from nilearn.surface import (
+    InMemoryMesh,
+    SurfaceImage,
+    load_surf_data,
+    load_surf_mesh,
+)
 from nilearn.surface.tests._testing import generate_surf
 
 try:
