+++ nilearn/surface/__init__.py
@@ -1,15 +1,14 @@
 """Functions for surface manipulation."""
 
-# TODO the following are not mentioned in the API part of the doc
-# "load_surface",
-# "check_surface",
-# "check_mesh_and_data",
-# "Mesh",
-# "Surface",
-
 from .surface import (
+    FileMesh,
+    InMemoryMesh,
     Mesh,
+    PolyData,
+    PolyMesh,
     Surface,
+    SurfaceImage,
+    SurfaceMesh,
     check_mesh_and_data,
     check_surface,
     load_surf_data,
@@ -19,8 +18,14 @@ from .surface import (
 )
 
 __all__ = [
+    "FileMesh",
+    "InMemoryMesh",
     "Mesh",
+    "PolyData",
+    "PolyMesh",
     "Surface",
+    "SurfaceImage",
+    "SurfaceMesh",
     "check_surface",
     "check_mesh_and_data",
     "load_surf_data",
