+++ nilearn/surface/surface.py
@@ -1,6 +1,8 @@
 """Functions for surface manipulation."""
 
+import abc
 import gzip
+import pathlib
 import warnings
 from collections import namedtuple
 from collections.abc import Mapping
@@ -16,6 +18,7 @@ from sklearn.exceptions import EfficiencyWarning
 
 from nilearn import _utils, datasets
 from nilearn._utils import stringify_path
+from nilearn._utils.niimg_conversions import check_niimg
 from nilearn._utils.path_finding import resolve_globbing
 from nilearn.image import get_data, load_img, resampling
 
@@ -1238,19 +1241,308 @@ def check_surface(surface):
     return Surface(mesh, data)
 
 
-def mesh_to_gifti(
-    coordinates,
-    faces,
-    gifti_file,
-):
+class PolyData:
+    """A collection of data arrays.
+
+    It is a shallow wrapper around the ``parts`` dictionary, which cannot be
+    empty and whose keys must be a subset of {"left", "right"}.
+
+    Parameters
+    ----------
+    left : :obj:`numpy.ndarray` or :obj:`str` or :obj:`pathlib.Path` or None,\
+           default = None
+
+    right : :obj:`numpy.ndarray` or :obj:`str` or :obj:`pathlib.Path` or None,\
+           default = None
+
+    Attributes
+    ----------
+    parts : :obj:`dict` of :obj:`numpy.ndarray`
+
+    shape : :obj:`tuple` of :obj:`int`
+    """
+
+    def __init__(self, left=None, right=None):
+        if left is None and right is None:
+            raise ValueError(
+                "Cannot create an empty PolyData. "
+                "Either left or right (or both) must be provided."
+            )
+
+        parts = {}
+        if left is not None:
+            if not isinstance(left, np.ndarray):
+                left = load_surf_data(left)
+            parts["left"] = left
+        if right is not None:
+            if not isinstance(right, np.ndarray):
+                right = load_surf_data(right)
+            parts["right"] = right
+
+        if len(parts) == 1:
+            self.parts = parts
+            self.shape = next(iter(self.parts.values())).shape
+            return
+
+        if parts["left"].shape[:-1] != parts["right"].shape[:-1]:
+            raise ValueError(
+                f"Data arrays for keys 'left' and 'right' "
+                "have incompatible shapes: "
+                f"{parts['left'].shape} and {parts['right'].shape}"
+            )
+
+        self.parts = parts
+        first_shape = next(iter(parts.values())).shape
+        concat_dim = sum(p.shape[-1] for p in parts.values())
+        self.shape = (*first_shape[:-1], concat_dim)
+
+    def to_filename(self, filename):
+        """Save data to gifti.
+
+        Parameters
+        ----------
+        filename : :obj:`str` or :obj:`pathlib.Path`
+                   If the filename contains `hemi-L`
+                   then only the left part of the mesh will be saved.
+                   If the filename contains `hemi-R`
+                   then only the right part of the mesh will be saved.
+                   If the filename contains neither of those,
+                   then `_hemi-L` and `_hemi-R`
+                   will be appended to the filename and both will be saved.
+        """
+        filename = _sanitize_filename(filename)
+
+        if "hemi-L" not in filename.stem and "hemi-R" not in filename.stem:
+            for hemi in ["L", "R"]:
+                self.to_filename(
+                    filename.with_stem(f"{filename.stem}_hemi-{hemi}")
+                )
+            return None
+
+        if "hemi-L" in filename.stem:
+            data = self.parts["left"]
+        if "hemi-R" in filename.stem:
+            data = self.parts["right"]
+
+        _data_to_gifti(data, filename)
+
+
+class SurfaceMesh(abc.ABC):
+    """A surface :term:`mesh` having vertex, \
+    coordinates and faces (triangles).
+
+    Attributes
+    ----------
+    n_vertices : int
+        number of vertices
+    """
+
+    n_vertices: int
+
+    # TODO those are properties are for compatibility with plot_surf_img.
+    # But they should probably become functions as they can take some time to
+    # return or even fail
+    coordinates: np.ndarray
+    faces: np.ndarray
+
+    def __repr__(self):
+        return (
+            f"<{self.__class__.__name__} "
+            f"with {getattr(self, 'n_vertices', '??')} vertices>"
+        )
+
+    def to_gifti(self, gifti_file):
+        """Write surface mesh to a Gifti file on disk.
+
+        Parameters
+        ----------
+        gifti_file : :obj:`str` or :obj:`pathlib.Path`
+            Filename to save the mesh to.
+        """
+        _mesh_to_gifti(self.coordinates, self.faces, gifti_file)
+
+
+class InMemoryMesh(SurfaceMesh):
+    """A surface mesh stored as in-memory numpy arrays.
+
+    Parameters
+    ----------
+    coordinates : :obj:`numpy.ndarray`
+
+    faces : :obj:`numpy.ndarray`
+
+    Attributes
+    ----------
+    n_vertices : int
+        number of vertices
+    """
+
+    n_vertices: int
+
+    coordinates: np.ndarray
+
+    faces: np.ndarray
+
+    def __init__(self, coordinates, faces):
+        self.coordinates = coordinates
+        self.faces = faces
+        self.n_vertices = coordinates.shape[0]
+
+
+class FileMesh(SurfaceMesh):
+    """A surface mesh stored in a Gifti or Freesurfer file.
+
+    Parameters
+    ----------
+    file_path : :obj:`str` or :obj:`pathlib.Path`
+            Filename to read mesh from.
+    """
+
+    n_vertices: int
+
+    file_path: pathlib.Path
+
+    def __init__(self, file_path):
+        self.file_path = pathlib.Path(file_path)
+        self.n_vertices = load_surf_mesh(self.file_path).coordinates.shape[0]
+
+    @property
+    def coordinates(self):
+        """Get x, y, z, values for each mesh vertex.
+
+        Returns
+        -------
+        :obj:`numpy.ndarray`
+        """
+        return load_surf_mesh(self.file_path).coordinates
+
+    @property
+    def faces(self):
+        """Get array of adjacent vertices.
+
+        Returns
+        -------
+        :obj:`numpy.ndarray`
+        """
+        return load_surf_mesh(self.file_path).faces
+
+    def loaded(self):
+        """Load surface mesh into memory.
+
+        Returns
+        -------
+        :obj:`nilearn.surface.InMemoryMesh`
+        """
+        loaded = load_surf_mesh(self.file_path)
+        return InMemoryMesh(loaded.coordinates, loaded.faces)
+
+
+class PolyMesh:
+    """A collection of meshes.
+
+    It is a shallow wrapper around the ``parts`` dictionary, which cannot be
+    empty and whose keys must be a subset of {"left", "right"}.
+
+    Parameters
+    ----------
+    left : :obj:`str` or :obj:`pathlib.Path` \
+                or :obj:`nilearn.surface.SurfaceMesh` or None, default=None
+            Mesh for the left hemisphere.
+
+    right : :obj:`str` or :obj:`pathlib.Path` \
+                or :obj:`nilearn.surface.SurfaceMesh` or None, default=None
+            Mesh for the right hemisphere.
+
+    Attributes
+    ----------
+    n_vertices : int
+        number of vertices
+    """
+
+    n_vertices: int
+
+    def __init__(self, left=None, right=None) -> None:
+        if left is None and right is None:
+            raise ValueError(
+                "Cannot create an empty PolyMesh. "
+                "Either left or right (or both) must be provided."
+            )
+
+        self.parts = {}
+        if left is not None:
+            if not isinstance(left, SurfaceMesh):
+                left = FileMesh(left).loaded()
+            self.parts["left"] = left
+        if right is not None:
+            if not isinstance(right, SurfaceMesh):
+                right = FileMesh(right).loaded()
+            self.parts["right"] = right
+
+        self.n_vertices = sum(p.n_vertices for p in self.parts.values())
+
+    def to_filename(self, filename):
+        """Save mesh to gifti.
+
+        Parameters
+        ----------
+        filename : :obj:`str` or :obj:`pathlib.Path`
+                   If the filename contains `hemi-L`
+                   then only the left part of the mesh will be saved.
+                   If the filename contains `hemi-R`
+                   then only the right part of the mesh will be saved.
+                   If the filename contains neither of those,
+                   then `_hemi-L` and `_hemi-R`
+                   will be appended to the filename and both will be saved.
+        """
+        filename = _sanitize_filename(filename)
+
+        if "hemi-L" not in filename.stem and "hemi-R" not in filename.stem:
+            for hemi in ["L", "R"]:
+                self.to_filename(
+                    filename.with_stem(f"{filename.stem}_hemi-{hemi}")
+                )
+            return None
+
+        if "hemi-L" in filename.stem:
+            mesh = self.parts["left"]
+        if "hemi-R" in filename.stem:
+            mesh = self.parts["right"]
+
+        mesh.to_gifti(filename)
+
+
+def _check_data_and_mesh_compat(mesh, data):
+    """Check that mesh and data have the same keys and that shapes match.
+
+    mesh: :obj:`nilearn.surface.PolyMesh`
+
+    data: :obj:`nilearn.surface.PolyData`
+    """
+    data_keys, mesh_keys = set(data.parts.keys()), set(mesh.parts.keys())
+    if data_keys != mesh_keys:
+        diff = data_keys.symmetric_difference(mesh_keys)
+        raise ValueError(
+            "Data and mesh do not have the same keys. "
+            f"Offending keys: {diff}"
+        )
+    for key in mesh_keys:
+        if data.parts[key].shape[-1] != mesh.parts[key].n_vertices:
+            raise ValueError(
+                f"Data shape does not match number of vertices for '{key}':\n"
+                f"- data shape: {data.parts[key].shape}\n"
+                f"- n vertices: {mesh.parts[key].n_vertices}"
+            )
+
+
+def _mesh_to_gifti(coordinates, faces, gifti_file):
     """Write surface mesh to gifti file on disk.
 
     Parameters
     ----------
-    coordinates : :class:`numpy.ndarray`
+    coordinates : :obj:`numpy.ndarray`
         a Numpy array containing the x-y-z coordinates of the mesh vertices
 
-    faces : :class:`numpy.ndarray`
+    faces : :obj:`numpy.ndarray`
         a Numpy array containing the indices (into coords) of the mesh faces.
 
     gifti_file: :obj:`str` or :obj:`pathlib.Path`
@@ -1269,13 +1561,12 @@ def mesh_to_gifti(
     gifti_img.to_filename(gifti_file)
 
 
-def data_to_gifti(data, gifti_file):
+def _data_to_gifti(data, gifti_file):
     """Save data from Polydata to a gifti file.
-
 
     Parameters
     ----------
-    data : :class:`numpy.ndarray`
+    data : :obj:`numpy.ndarray`
         The data will be cast to np.uint8, np.int32) or np.float32
         as only the following are 'supported' for now:
         - NIFTI_TYPE_UINT8
@@ -1304,3 +1595,162 @@ def data_to_gifti(data, gifti_file):
 
     gii = gifti.GiftiImage(darrays=[darray])
     gii.to_filename(Path(gifti_file))
+
+
+def _sanitize_filename(filename):
+    """Check filenames to write gifti.
+
+    - add suffix .gii if missing
+    - make sure that there is only one hemi entity in the filename
+
+    Parameters
+    ----------
+    filename : :obj:`str` or :obj:`pathlib.Path`
+        filename to check
+
+    Returns
+    -------
+    :obj:`pathlib.Path`
+    """
+    filename = Path(filename)
+
+    if not filename.suffix:
+        filename = filename.with_suffix(".gii")
+    if filename.suffix != ".gii":
+        raise ValueError(
+            "Mesh / Data should be saved as gifti files "
+            "with the extension '.gii'.\n"
+            f"Got '{filename.suffix}'."
+        )
+
+    if "hemi-L" in filename.stem and "hemi-R" in filename.stem:
+        raise ValueError(
+            "'filename' cannot contain both "
+            "'hemi-L' and 'hemi-R'. \n"
+            f"Got: {filename}"
+        )
+    return filename
+
+
+class SurfaceImage:
+    """Surface image containing meshes & data for both hemispheres.
+
+    Parameters
+    ----------
+    mesh : :obj:`nilearn.surface.PolyMesh`, \
+           or :obj:`dict` of  \
+           :obj:`nilearn.surface.SurfaceMesh`, \
+           :obj:`str`, \
+           :obj:`pathlib.Path`
+           Meshes for the both hemispheres.
+
+    data : :obj:`nilearn.surface.PolyData`, \
+           or :obj:`dict` of  \
+           :obj:`numpy.ndarray`, \
+           :obj:`str`, \
+           :obj:`pathlib.Path`
+           Data for the both hemispheres.
+
+    Attributes
+    ----------
+    shape : (int, int)
+        shape of the surface data array
+    """
+
+    def __init__(self, mesh, data):
+        """Create a SurfaceImage instance."""
+        self.mesh = mesh if isinstance(mesh, PolyMesh) else PolyMesh(**mesh)
+
+        if not isinstance(data, (PolyData, dict)):
+            raise TypeError(
+                "'data' must be one of"
+                "[PolyData, dict].\n"
+                f"Got {type(data)}"
+            )
+
+        if isinstance(data, PolyData):
+            self.data = data
+        elif isinstance(data, dict):
+            self.data = PolyData(**data)
+
+        _check_data_and_mesh_compat(self.mesh, self.data)
+
+        self.shape = self.data.shape
+
+    @classmethod
+    def from_volume(
+        cls, mesh, volume_img, inner_mesh=None, **vol_to_surf_kwargs
+    ):
+        """Create surface image from volume image.
+
+        Parameters
+        ----------
+        mesh : :obj:`nilearn.surface.PolyMesh` \
+             or :obj:`dict` of  \
+             :obj:`nilearn.surface.SurfaceMesh`, \
+             :obj:`str`, \
+             :obj:`pathlib.Path`
+             Surface mesh.
+
+        volume_img : Niimg-like object
+            3D or 4D volume image to project to the surface mesh.
+
+        inner_mesh : :obj:`nilearn.surface.PolyMesh` \
+             or :obj:`dict` of  \
+             :obj:`nilearn.surface.SurfaceMesh`, \
+             :obj:`str`, \
+             :obj:`pathlib.Path`, default=None
+            Inner mesh to pass to :func:`nilearn.surface.vol_to_surf`.
+
+        vol_to_surf_kwargs : dict[str, Any]
+            Dictionary of extra key-words arguments to pass
+            to :func:`nilearn.surface.vol_to_surf`.
+
+        Examples
+        --------
+        >>> from nilearn.surface import SurfaceImage
+        >>> from nilearn.experimental.surface import load_fsaverage
+        >>> from nilearn.datasets import load_sample_motor_activation_image
+
+        >>> fsavg = load_fsaverage()
+        >>> vol_img = load_sample_motor_activation_image()
+        >>> img = SurfaceImage.from_volume(fsavg["white_matter"], vol_img)
+        >>> img
+        <SurfaceImage (20484,)>
+        >>> img = SurfaceImage.from_volume(
+        ...     fsavg["white_matter"], vol_img, inner_mesh=fsavg["pial"]
+        ... )
+        >>> img
+        <SurfaceImage (20484,)>
+        """
+        mesh = mesh if isinstance(mesh, PolyMesh) else PolyMesh(**mesh)
+        if inner_mesh is not None:
+            inner_mesh = (
+                inner_mesh
+                if isinstance(inner_mesh, PolyMesh)
+                else PolyMesh(**inner_mesh)
+            )
+            left_kwargs = {"inner_mesh": inner_mesh.parts["left"]}
+            right_kwargs = {"inner_mesh": inner_mesh.parts["right"]}
+        else:
+            left_kwargs, right_kwargs = {}, {}
+
+        if isinstance(volume_img, (str, Path)):
+            volume_img = check_niimg(volume_img)
+
+        texture_left = vol_to_surf(
+            volume_img, mesh.parts["left"], **vol_to_surf_kwargs, **left_kwargs
+        )
+        texture_right = vol_to_surf(
+            volume_img,
+            mesh.parts["right"],
+            **vol_to_surf_kwargs,
+            **right_kwargs,
+        )
+
+        data = PolyData(left=texture_left.T, right=texture_right.T)
+
+        return cls(mesh=mesh, data=data)
+
+    def __repr__(self) -> str:
+        return f"<{self.__class__.__name__} {getattr(self, 'shape', '')}>"
